Repair OBS -> UpDeck Buttons
In prior versions, there was a feature that allowed the user to bind a
keycode in OBS that would send to UP Deck to be read and executed.

This was broken due to a change that introduced decks, previously it
looked like buttons where individual files inside of a single deck.

We repaired this by piggy-backing on a not-intented interface to send
the command back to the UPDeck Server. This results in a less than
ideal delay of up to 8 seconds (the update interval in the client).

The delay is due to the fact that this buffer will only be read if a
command is sent by the actual client as there is no direct pulling for
this queue. Fortunately the client sends a update to read any of the
volumes (configured or not) every 8 seconds.

To use this, you need a) have a deck saved on the server b) you need
to set this deck name in the OBS Script settings